mem: frame: Define dealloc() signature
Prepare the `FrameAllocator` for deallocation by defining the signature
for the `dealloc` method.

The previous function signature was a placeholder. To properly implement
frame deallocation, the method requires mutable access to the allocator
itself (`&mut self`) to modify its free list. It also needs to accept
the frame or list of frames to be freed.

This change introduces the `head_frame` parameter of type
`NonNull<Frame>`, which will represent the head of a linked list of
frames to be returned to the allocator. The implementation of the
deallocation logic will follow in a subsequent commit.